Add tests for SolverClientOrders
